refactor out the messages.ts
